Move parseCsv to shared
